Add anysize
This patch allows st to resize to any pixel size, makes the inner
border size dynamic, and centers the content of the terminal so that
the left/right and top/bottom borders are balanced. With this patch,
st on a tiling WM will always fill the entire space allocated to it.

https://st.suckless.org/patches/anysize/st-anysize-0.8.1.diff